Conditionally enable PCI in hwloc
Co-locale support for multiple NICs and GPUs requires PCI support in hwloc.
Previously, PCI support was only enabled if CHPL_COMM==ofi, which was tested
in the third-party/hwloc/Makefile. Unfortunately, this meant it wasn't
embedded in the build path, so that changing CHPL_COMM required manually
rebuilding hwloc, which was obviously error-prone. Now PCI enable/disable is
controlled by the CHPL_HWLOC_PCI variable which is embedded in the build
paths for both the runtime and third-party/hwloc, and whose default value is
determined by CHPL_COMM, CHPL_LOCALE_MODEL, and CHPL_GPU.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>