Improve the error encountered when accidentally making an iterator initializer
Prior to this change, users could write `iter init` and would only discover
their error when they tried to initialize the type.  However, they wouldn't
discover the error with a nice error message - instead, they'd get an internal
compiler error from function resolution because we didn't actually expect anyone
to make that mistake and so didn't appropriately protect ourselves from it.

This change adds a check after parsing to make sure that the user doesn't try to
make an iterator initializer and gives an appropriate error message if so.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>