Add two tests locking in the error message added in the previous commit
The first is the code reported by the user who encountered this situation in the
first place - this code did not encounter the internal compiler error, but was
still wrong.

The second actually tries to initialize the class and so previously encountered
the internal compiler error.

Both tests now appropriately give the error message.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>