Allow scope lookups to traverse beyond compiler-generated modules
Modifies the idToParentId map of a BuilderResult to point to the module
ID from which the compiler-generated module was created.

In the scope query logic, we then reject a compiler-generated module as
a boundary for lookups, and allow the logic to proceed as though the
compiler-generated module lived within the original code.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>