Compute and store param enum element name when created
Prior to this change, param enums could not be stringified because
EnumParam only stored an ID and the stringify signature lacks a
``Context*`` needed to map the ID to uAST. Without the uAST we cannot
compute the original name of the enum element.

Instead, store a pair of ``ID`` and ``string`` inside EnumParam and
compute the element's name ahead of time.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>