Unstable anon script special case <command line>
This adds a few things to the unstable warning anonymizer script in
order to update it for the 2.2 release.

- Add handling for a warning starting with the word `<command line>`:
  previously, we were assuming that a warning either starts with a
Chapel file name (like `foo.chpl`) or with `<command-line arg>` but a
new unstable warning was added this release which starts with `<command
line>` and it also doesn't have a number following it (the other two
have numbers following them which indicate a line number of argument
number)
- Add special case handling for the ambiguous modules warning (which is
  what also required the change from the 1st bullet) since it exposes
implementation details (module names), we scrub these names now.
- Add testing for the special case

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>