Ensure variable declarations at the end of try blocks are cleaned up
My reading is that the error handler in desugared try-catch (which
is just 'if error, forward it' is traversed early, before the last
variable in the `try` is not marked as initialized, so that auto-destroy
calls are not inserted for the (potentially uninitialized) variable.

However, when the early handling occurs, the "insert auto-destroys
after last mention" logic fires for the if-error conditional,
and marks all variables as having been deinitialized. This isn't
accurate, since that one variable not marked as initialized gets
marked deinited, even though it is not cleaned up (since it was
not init'ed).

The solution is not to insert last-mention auto-destroys when
doing the early traversal of the if-error code. Fortunately,
it looks like the if-error stmt is traversed a second time, in
normal order, which happens _after_ the last variable is
marked initialized. The last-mention logic can be executed there,
and will properly clean up variables as needed.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>