Add support for qualified interface access
Strangely, even though we have an `isym` when building
the class hierarchy, I chose to use the symbol's name
to construct an `implements` statement. This works when
the name is available in scope without qualification, but
not when qualified access is used. The principled solution
is to allow interface statements to be constructed using
a known interface symbol. This PR does that, and switches
the class hierarchy logic to avoid using a name.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>