Act on part of Engin's feedback
The main piece of feedback I'm acting on here is to update the comment
in StencilDom.doiLocalStoredSubdomains(), which had fallen out of date
when I removed the 'locale' argument to the iterator in a
late-breaking commit.  Here, I've updated it to note that the current
call should be communication-free since we are always answering the
question for 'here'.  I've also updated the stencil localAccess.chpl
test to wrap the localAccess calls in a `local` block to lock in that
no communication is required to implement the calls.

Then I took that same principle to the localAccess-block*.chpl tests
to lock in that the block localAccess calls do not require
communication.

Note that I used an unusual formatting scheme here because I didn't
want to have to update the line numbers of the error messages... :(

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>