Add illegal local access back in to capture error
Engin pointed out, rightly, that in updating these tests to make the
localaccess calls legal, I aso pointed out that we had a pre-existing
bug specific to the case of oversubsctibed locales that we didn't have
any tests catching.  So here, I've added the illegal accesses back at
the end of the test to lock in the remote accesses.

The reason that these cases are different is that our dsiLocalAccess()
calls on Block and Cyclic use different (arguably incorrect) logic
when used in an oversubscribed manner.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>