Update stencil oversubscription test for multiple locales
While reviewing the block and cyclic versions of this test, it wasn't
clear to me why this stencil version of the test didn't require changes
when the others did.  It turns out that the answer is that it wasn't
ever enabled for multi-locale runs...  Fixing that here, and updating
it similar to the block and stencil versions.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>