Improve initializers on classes that inherit
This commit improves support for initializers on classes that inherit.
With this commit dyno now supports:
- implicit 'super.init' calls
- errors for accessing parent fields before 'super.init'
- usage of initialized fields part-way through initializing the whole type

This commit updates the Resolver to rely somewhat less on 'InitResolver'
for explicitly handling ``Identifier`` and ``Dot`` uAST nodes. Instead
we continue to rely on established Resolver mechanisms and selectively
invoke the ``InitResolver`` when appropriate.

This commit also improves a fix within 'handleCallExpr' originally meant
to support receiver-less 'init' calls. Instead, this fix is adapted to
any 'init' call within an initializer, and will replace or create the
receiver using a fully-generic type. This is done in case the user has
erroneously initialized fields prior to the explicit 'init' call.

------------------

The ``fieldIdFromPossibleMentionOfField`` helper method is updated to
return a field ID (if found) and whether that field is a parent field.

Some logic from ``handleCallToSuperInit`` is split off into a new
``updateSuperType`` method to better support the implicit 'super.init'
case.

Minor fixes to ``findFieldByName``
- Use ``VarLikeDecl`` instead, as ``NamedDecl`` can match primary methods.
- Do not continue searching upwards if the parent class is the root class.

Lastly, this commit adds tests for newly supported initializer behavior

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>