Generate uAST for default initializers
This commit replaces the current signature-based implementation for
default initializers with uAST for a Function. This Function's body
contains the suitable statements to initialize each field, as well as a
"super.init" call if applicable. This code currently lacks suitable
init-expressions for default-initializable formals, which is left as
future work.

An UntypedFnSignature is built manually so that we can force resolution
to accept non-existent default values for each formal, regardless of
whether it has an init-expression or not.

The resulting TypedFnSignature is created with `typedSignatureInitial`.
This might be overkill, but it keeps the code cleaner for the time
being.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>