Use a new Mason environment variable instead of a set location
Adds `MASON_LICENSE_CACHE_PATH` as something that Mason understands, and when it
is specified, use it with `--reference-if-able` when downloading the license
repository.

Updates the test EXECENV to set this environment variable to REPO_CACHE_PATH,
which we use for other cached repositories in nightly testing, but only if that
environment variable is defined.  This should allow fellow team members to not
have to set this environment variable (because they probably won't need it)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>