Fix method resolution on string and bytes types
Updates CompositeType to use the leading-underscore version of
string, bytes, and locale types so that they match the records declared
in module code. The 'is[String|Bytes|Locale]Type' methods are updated to
instead check the ID of the type, rather than the name. Lastly,
stringify is updated to print the proper reserved names for these types.

Note: The 'locale' type is not tested here due to a pre-existing
recursive query bug.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>