Fix erroneous ambiguity error for nested functions
Logic in ``resolveIdentifier`` was looking for ambiguity errors between
parenless methods and locally-declared variables, but was instead
issuing errors for paren-ful methods and nested functions as well.

This commit changes resolveIdentifier to only run these errors if a
parenless method is found and if there are redeclarations other than the
paren-less method.

``test15`` in ``testMethodCalls`` is temporarily retired as it is not
actually considered to be an ambiguity in production. This instead
appears to be a bug in which we incorrectly prefer the method over the
nested function.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>