Improve generation of super.init() call in default initializers
This commit fixes an issue where super.init() calls were being generated
despite the lack of a parent class. This lead to resolution failures
when trying to resolve init on the root class.

This kind of error manifested itself when there was an interface in the
inherit-exprs list of the class. The 'initHelper' function was naively
ignoring the possibility of such expressions.

The bug is fixed by first relying on ``initialTypeForTypeDecl`` to
compute the parent class rather than doing it manually inside
'initHelper'.

-----

While here, it was convenient to fix another super.init() related bug
when generating default initializers when a parent (or other ancestor
class) had a user-defined initializer. In such cases the compiler should
not be generating formals for classes that define their own initializer,
and should instead be invoking an empty 'super.init()' when such a class
is the direct parent.

A straightforward way to fix this issue was to abandon the recursive
nature of the 'initHelper' function and instead build the default
initializer for the parent class. With a correct parent initializer we
can simply copy the formals to the child class initializer and avoid
having to think recursively.

Lastly, 'initHelper' is modified to return a boolean indicating whether
a 'super.init()' call is needed, as opposed to examining the
inherit-exprs of the class.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>