At Michael's suggestion, limit the application to `PRIM_GET_MEMBER_VALUE`
Michael pointed out that `PRIM_GET_MEMBER_VALUE` can situationally return a ref
or a value and we don't want to apply the const-ness in the value case (as it
will be a copy and thus independent of the behavior of the `const` source).  So
ensure that the result of the call will be a ref before propagating the `const`.

This passed the normally impacted tests locally, will run a full paratest as
well

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>