Make CHPL_LIB_PATH ignore empty strings
If CHPL_LIB_PATH has colons in it, we split the path based on the
colons. However, if we have a trailing or leading colon, or two colons
back to back (`::`) in CHPL_LIB_PATH, it get's added as an empty string.
This cases errors since there isn't an argument to go with `-L` for the
target compiler.

This makes the change such the the `addPath` function ignores the empty
strings in CHPL_LIB_PATH.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>