c2chapel dependency changes
Undo version upgrade for pycparser because it is not actually compatible with
the newest version of pycparserext (despite it apparently installing cleanly if
you didn't have an older version of pycparserext installed on your machine, so
far as I can tell)

Explicitly set the version of pycparserext to the latest version.  This should
help avoid the testing mishap in the future, since explicitly specifying the
version makes it tell us if we upgrade to a version of pycparser that is
incompatible with it.

Update the Makefile for c2chapel to stay in line with the pycparser version

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>