Updates and fixes to task parallel primer
This updates the task parallel primer to use the term 'task' instead
of thread.  While spot-checking the sections affected by this, I also
made some localized changes to improve the clarity and precision of
the descriptions.  I also added some more material at the front to
define what a 'task' is (roughly) and to move a `config const` down
closer to where it's used.

I also noticed that the messages printed by the main/first coforall
example in this primer were very weird, implying that each statement
within the body was printed by a distinct task.  This may have been a
copy-paste error from the second coforall loop example, which uses
'begin's, but still surprises me that it'd been around for so long.

Note that I did not do a complete read-through of the primer in the
interest of time, but given the above, think it may be beneficial
to do so.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>