Improve querying of generated uAST
Previously, generated uAST was being created given a particular type
possibly with other information about how the call was being invoked.
Generated uAST was then created as a fake 'included' module with a
parent symbol path was that of the module in which the type was defined.
This uAST was then stored and retrieved using some set/get queries in
``chpl::parsing``.

This commit updates the implementation to instead rely on generated
functions using the type's symbol path as its parent path. This allows
``parseFileContainingIdToBuilderResult`` to use a generated ID's symbol
path and symbol name to query the uAST, as opposed to using the old
getter method. Now, we rely on the ``buildDefaultFunction`` query in
``chpl::resolution``.

For example, a default initializer will have a path like
``MyMod.MyType.init``. With the parent symbol path of ``MyMod.MyType``
we can query the type's uAST if needed, and with the name ``init`` we
know which uAST-generating function to invoke. This allows us to get rid
of the clunky 'chpl__generated' prefix to generated modules. We also now
need to avoid generating an implicit module for generated uAST and
instead have a Function as the top-level expression in a Builder.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>