Refactor 'buildTypeConstructor' for new generated uAST query system
This commit updates buildTypeConstructor to build a type constructor
given only a type's ID.

Now that the untyped signature is not pointing to non-existent uAST, we
can also stop using the original type's ID in the untyped signature.
This requires some minor modifications to return-type-inference now that
'idIsFunction' is returning true for type constructors.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>