Clean up 'generatedFrom' usage in Builder/BuilderResult
This commit updates Builder to store a boolean indicating whether the
uAST is generated rather than storing the old 'generatedFrom' ID. The
necessary information from this ID can and should be stored in the
pre-existing ``startingSymbolPath_`` field.

After some experimentation it turns out that BuilderResult no longer
needs a ``generatedFrom`` ID to use for the id-to-parent map. This is
because we are now correctly setting the parent path for generated IDs
in ``parseFileContainingIdToBuilderResult``.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>