Only clean up the chapel library runtime if we initialized it to begin with
This solves an error I was seeing with Python interoperability.  When a
generated library was in the PYTHONPATH but not initialized, we were seeing this
error message:

```
runtime/include/tasks/qthreads/chpl-tasks-impl-fns.h:140: chpl_task_getInfoChapel: Assertion `data && data->bundle' failed.
```

This seemed to be because the generated `__init__.py` file is called in more
circumstances than just if the module itself was explicitly imported into a
program, meaning that the `atexit` trigger we added to cleanup the library was
firing, despite us never having called `chpl_setup()`.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>