Clarify assumptions for parents of generated IDs
Firstly, fixes a bug in parentSymbolId where getting the parentSymbolId
of a non-scope-defining symbol (e.g. an Identifier) would incorrectly
return a generated ID with the grandparent symbol path. This does not
seem to be exposed in any testing at this time.

Adds some comments explaining the assumption that there is only one
scope-defining symbol in generated uAST. Also adds an assertion to
enforce this rule, which was convenient to add in doAssignIds.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>