Treat passing to unknown type as instantiation
This is important for cases in which type queries need to be populated.
They are only populated from substitutions, but substitutions
are not created for completely unknown formals, because
prior to this commit, a completely unknown formal can be
passed to without instantiating (i.e., without substitution).
This commit changes that.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>