Reject candidates that fail due to inability to infer type queries
After computing a formal type, we traverse the formal AST to find
out what the type queries are. This may fail (e.g., if the formal
AST doesn't match up with the instantiated type). Thus, the post-
query type will be unknown. This is an error, since it indicates
that the type queries couldn't be properly made sense of and
incorporated into the final type info. In such cases, reject the
candidate.

A concrete example: formal ?k*?t, actual int. The initial type
is unknown; after instantiation, the formal type is int. When
resolving queries, we can't figure out how to spit init across
?k*?t, leaving type queries unknown. When re-combining the formal
type with query info, ?k*?t is again unknown since the queries
were not known. This means we couldn't instantiate ?k*?t with int,
and the candidate is rejected.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>