Handle unknown actuals (needed for out-formals) in builtin functions
We have some builtin functions like type/param casting to string,
handled by the compiler. They did not expect `null` actual types,
but this can happen sometimes (e.g., when the call site allows
for the possibility of `out`-formals being matched with the actual).
This commit adjusts the code to handle that.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>