Skip unknown arguments when building ranges.
Resolving ranges boils down to resolving a function call. Dyno
skips resolving normal function calls under certain conditions
(e.g., if the argument types could not be inferred). However,
it doesn't do so for ranges, which made it possible for 'Unknown'
or otherwise invalid actuals to slip into the call resolution
process. This commit adjusts the resolver to re-use the skipping
logic.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>