Back out one of the changes in #26178 to get warnings/warnings.chpl working
While paratesting my branch tonight, I found that

  chplenv/chplconfig/warnings/warnings.chpl

was failing due to PR #26178.  Though I'm not that familiar with this
logic, it seems that reverting half of the changes in it keeps this
test working as expected, though it may suggest we need to do more to
both achieve the PR's aim and also keep errors for cases like this.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>