Make the python interoperability jobs fail if dependencies are not installed
Adds explicit testing of the presence of Cython and numpy to both jobs, and
ZMQ to the multilocale job.

Ensures that the scripts fail if these dependencies are not found.

Does this by creating a couple of helper files:
- a common bash script for checking Cython and numpy
- a python script for getting access to our compiler information.  This seemed
easier than trying to figure out how to do it in bash

When checking for ZMQ, create a temporary directory with a C file that checks if
the header is appropriately installed.  This is similar to the skipif handling
for the job, but distinct due to running before the Chapel compiler is built
(and thus before we can use the `compileline` command like that skipif does).  I
could have potentially updated the skipifs to use this pathway, but left that
as future work

This behaved appropriately on my local system (where ZMQ is not installed).
Checking its behavior on a linux machine where ZMQ is not installed as well

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>