Correctly compute field types in initializers when types do not match
Prior to this PR InitResolver would always use the RHS type of field
initialization statements to set the type of the field. For example,
in "this.foo = RHS;", the field 'foo' would always use the type of RHS
when resolving any relevant initialization calls (e.g. assignment).

This PR uses Resolver::getTypeForDecl to check that the types are
compatible, and then uses the computed type for 'foo'.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>