Allow resolution of generic init= with concrete formals
Consider the following program:

```chpl
record G { type T; var x : T; }

proc G.init=(other: this.type) {
  this.T = other.T;
  this.x = other.x;
}

proc G.init=(other: int) {
  this.T = int;
  this.x = other;
}

var x : G(?) = 5;
```

Previously this program would fail to compile, and give a confusing
error message stating it had found the second 'init=' specifically for
'int', but that it somehow wasn't a candidate.

The bug was caused by some missing logic in ``computeSubstitutions``
which was incorrectly returning 'false' in this case. The solution was
to manually ignore a generic 'this' formal for initializers.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>