Jade pointed out I ran afoul of a known syntactic limitation, so update tests
`proc (` is not supported due to parser conflicts, only `proc(`.  Updated the
tests with that in mind.  The assignInIfExpr test still encountered a syntax
error, so update the .bad file, and point the .future file at the original
issue.

The assignLater test also needed to update the syntax declaring the type of the
variable, but after that the test worked like the original version did at the
time it was created, so removed the .bad and .future file.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>