Optimize transfers of sparse CSR/CSC arrays with --no-bounds-checks/--fast
This fixes an oversight from #25243 in which I applied an optimization
to DefaultSparse layouts that I failed to also replicate for CSR/CSC.
The issue that we would do a check of domain equivalence prior to
bulk-copying a remote sparse array, even if bounds checking was off,
and this resulted in O(nnz) extra work, including O(nnz) communication
if the source and destination arrays match.  This updates the relevant
bulk-transfer helper routines in the CSR/CSC layout to match the logic
of DefaultSparse, only doing the check if checks are on.  It also adds
in a test to lock in these comm counts with and without --fast.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>