Incorporate Matt's suggestion for the subprocess
I meant to include `python3 -c` in the command when transforming it from a call
in the body of the Python program to a subprocess call but missed it (and
somehow didn't run into problems doing so).  Still, it's best to be explicit
here in case it makes a difference.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>