Fix the suppressif file to actually detect the failure condition
Matt pointed out that I had accidentally written the subprocess command in such
a way that subprocess was escaping the whole command and so acting like it
detected the failure mode when really it was just not running the command in a
way that would work.  After fixing that, I also needed to hide the failure
output from start_test so it wouldn't think the suppressif had failed.  With
that all fixed, we now appropriately suppress the failure mode

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>