Add a new primitive to fix the type of the original variable
The type of the temporary got updated with the previous commit, but then we were
encountering an issue where the later use of the variables (in the writeln) was
still failing to have an actual type.  This commit allows us to update the type,
though it still doesn't allow us to compile the program - but the error message
for the user actually points to the problem and so leaves us in a better state
than the internal compiler error.

To do, add a new primitive, and insert that primitive when we create the addrOf
call for the variable.  This allows us to let the temporary get updated and then
check back in for its type prior to any later uses of the variable.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>