Fix a bug I'd overlooked
In reviewing my PR, Lydia noticed something I hadn't: That the
coordinates array had some extra entries at the end, due to the fact
that we use recursive doubling on the nnzDom domain such that it isn't
always exactly nnz elements large.  Here, I've addressed that by
resizing the domain to be just the right size before returning the
references to the coordinates.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>