Improve error when using -j with LLVM back-end
In issue #26326, Engin points out that the historical wording wasn't
all that useful for users using '-j' who might not have any idea what
the '--incremental' flag does.  This generalizes the error to mention
those flags as well, and updates an existing test to test that it
fires in all three cases and to reflect the new error message.

Resolves #26326.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>