Remove the warning for changing default parSafe mode for associative domains
Associative domains with parSafe=false were stabilized in #24352
We warned about the change in the default parSafe mode, now that we have had a few releases since that, we can go ahead and get rid of that warning and all the related logic.

This PR removes the special warning for using the default parSafe when using associative domains. We leave the warning for parSafe=true as it is since that is still unstable.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>