Update docs to use 'init this;' rather than 'this.complete();'
While reading up on the status of throwing initializers, I was
surprised to find the technote still refers to 'this.complete();'
rather than 'init this;'.  This PR primarily updates that
documentation, though I also looked for other cases that might need
updating to reduce the number of hits when 'git grepping' for the
expression.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>