Close leak in throwing postinits
OK, well this is embarrassing.  Despite asserting in the PR message of
inits in terms of what it did, apparently I never actually ran with
memleaks on to make sure that no memory was leaked when a postinit
throws, and...  Well, it was.  (I had inspected the generated code and
seen that we were emitting some code to handle its thrown errors, but
didn't look closely enough to see that that handling didn't actually
do any clean-up).  Sloppy.

Here, I'm fixing my issue by tucking the call to postinit() into the
same try...catch block as the throwing initializer code is put into,
which gives it the same outcome, which is to free the memory.  This
should resolve the memory leaks failures we saw on Dec 20, 2024.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>