Register 'as' clauses in manage statements when converting.
Previously, we didn't correctly find them during Dyno scope resolution,
and later the production scope resolver took care of them. However, now
that we find them, we have to note them as converted symbols so that
when a toId points at an 'as' Variable, we know what to point the
SymExpr to.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>