Snapshot of fixing Set zippering
Adds a pair of leader/follower iterators to ChapelHashtable and a helper method,
and calls them from the Set `these` iterators.

The implementation does frequent recomputation of how to divide up the iteration
space to find full slots.  This can probably be improved on, but I wanted to
take a snapshot of what worked before attempting it.

It also does not help when zippering with arrays yet.  That is also a next step

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>