Get the set/hashtable iterator to work with arrays as well
This meant always computing the number of chunks and which chunk we were
ourselves, since the array iterator expects the other components of the tuple to
be ranges as well (so we can't use followThis to send that information) and the
argument list for the follower is supposed to match the argument list of the
leader aside from followThis.  This is unfortunate, but given the sizes I
checked, seemed to be okay.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>