Fix temporary file cleanup
The command buffers used to delete the PBS temporary files were too small,
causing the file names to be truncated and the files to not be deleted. The
launcher now unlinks the files, obviating the need for a command buffer. This
approach also generates an error message if the unlink fails.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>