Improve the error message for Set's part of the iterators
Use a halt wrapper instead of `__primitive("chpl_error"`, especially because
this is a new error instead of trying to match errors that the compiler can
insert.

While doing that, made the error messages clearer about what was wrong and what
should be done instead (and hopefully head off confusion for the follower case).
And add a test locking in the error message in the leader case

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>