Fix a few testing errors with promotedSizeMismatch
First, I made the classic mistake of failing to think about --fast
testing, which turns off the bounds checking this test is looking
for.  Here, I add --checks to its .compopts file to try and avoid
this going forward.

More subtly, it didn't occur to me that the test has a race since the
leader will fire off multiple follower tasks, giving the potential to
have the size mismatch in dimension 0 or 1 to appear first.  Here, I
side-stepped that by only making a single dimension mismatch in size.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>