Simplify the implementation, at Brad's suggestion
Brad pointed out that we could just count through the filled elements of the
hashtable until we reach the count specified by the `low` value, and then start
yielding.  This is more performant than what I had been doing and it allows more
flexibility, enabling it to actually work with ranges and arrays

The one drawback is that it does mean we don't detect when the leader is shorter
than the follower, but that's already true for arrays so we'll at least be
consistent with them.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>