Update docs to point to new webpages rather than old
This updates URLs within our docs to point to the new website rather
than the URLs from the old website.  I believe most of these cases
would automatically redirect to the new website, but the more we
establish the new URLs as the canonical versions, the happier our
Google dashboard will be.

The one change here that's a little borderline is
https://chapel-lang.org/bugs.html which has been a (client-side)
redirect for ages, though what it redirects to isn't the greatest of
pages.  Here, I've updated one such instance to avoid the redirect,
but there are lots of others that I skipped over, feeling like we
should probably establish a better bugs page in general.  I've opened
an issue on the chapel-www repository to determine what we should do
here.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>